初期描画のカスタムフック
